RankingsTest

+++ app/Http/Controllers/GameController.php
@@ -14,6 +14,9 @@ class GameController extends Controller
     {
         $user = User::find($id);
 
+        if (!$user) {
+            return response()->json(['message' => 'Jugador no trobat'], 404);
+        }
         $dice1 = random_int(1, 6);
         $dice2 = random_int(1, 6);
 
@@ -33,6 +36,9 @@ class GameController extends Controller
     public function deleteGame($id)
     {
         $user = User::find($id);
+        if (!$user) {
+            return response()->json(['message' => 'Jugador no trobat'], 404);
+        }
 
         $user->games()->delete();
